requirements.txt
# AI Card Trader - Python Requirements
# Run: pip install -r requirements.txt

# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and vector store
qdrant-client==1.6.9
sentence-transformers==2.2.2
transformers==4.30.2
huggingface_hub==0.16.4

# LLM integration
requests==2.31.0
httpx==0.25.2

# Data handling
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0

# Frontend options
streamlit==1.28.1
gradio==4.7.1

# Utilities
python-dotenv==1.0.0
uuid==1.30
asyncio-mqtt==0.13.0
tqdm==4.66.1

# Optional: Direct Ollama Python client
ollama==0.1.7

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
